+++ src/parser.tab.c
@@ -165,7 +165,7 @@ typedef union YYSTYPE
 /* Line 214 of yacc.c  */
 #line 20 "parser.y"
 
-    long double fval;
+    double fval;
     char *str;
     ASTNode *ast;
     VarType vtype;
